unify completion post composer
