+++ src/components/Quest.tsx
@@ -1,29 +1,17 @@
-import React, { useRef, useState } from "react";
-import {
-  Animated,
-  Image,
-  KeyboardAvoidingView,
-  Modal,
-  Platform,
-  StyleSheet,
-  TextInput,
-  TouchableOpacity,
-  View,
-} from "react-native";
+import React from "react";
+import { StyleSheet, View } from "react-native";
 import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
-import MaterialIcons from "@expo/vector-icons/MaterialIcons";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useAuth } from "../contexts/AuthContext";
 import { useLanguage } from "../contexts/LanguageContext";
 import { colors } from "../constants/Colors";
 import { Text } from "./StyledText";
 import { SideQuest } from "../types/sideQuests";
-import { useMediaUpload } from "../hooks/useMediaUpload";
-import { Loader } from "./Loader";
 import { captureEvent } from "../lib/posthog";
 import { SIDE_QUEST_EVENTS, USER_PROPERTIES } from "../constants/analyticsEvents";
 import { setUserProperties } from "../lib/posthog";
 import { supabase } from "../lib/supabase";
+import { CompletionPostComposer } from "./CompletionPostComposer";
 
 interface QuestCardProps {
   quest: SideQuest;
@@ -83,12 +71,8 @@ export const QuestCard: React.FC<QuestCardProps> = ({ quest, tags }) => {
 };
 
 export const ShareQuestExperience: React.FC<ShareQuestExperienceProps> = ({ quest }) => {
-  const { t } = useLanguage();
   const { user } = useAuth();
   const queryClient = useQueryClient();
-  const [modalVisible, setModalVisible] = useState(false);
-  const [comfortRating, setComfortRating] = useState(3);
-  const fadeAnim = useRef(new Animated.Value(0)).current;
 
   const questCompletionQuery = useQuery({
     queryKey: ["quest-completion", quest.id, user?.id],
@@ -110,266 +94,33 @@ export const ShareQuestExperience: React.FC<ShareQuestExperienceProps> = ({ ques
   const completed = questCompletionQuery.data ?? false;
   const checking = questCompletionQuery.isLoading;
 
-  const {
-    selectedMedia,
-    postText,
-    setPostText,
-    isUploading,
-    isSubmitting,
-    uploadProgress,
-    handleMediaUpload,
-    handleRemoveMedia,
-    handleSubmit,
-  } = useMediaUpload({
-    onUploadComplete: () => {
-      setModalVisible(false);
-      queryClient.setQueryData(["quest-completion", quest.id, user?.id], true);
-      queryClient.invalidateQueries({ queryKey: ["home-posts"] });
-      captureEvent(SIDE_QUEST_EVENTS.DAILY_QUEST_COMPLETED, {
-        quest_id: quest.id,
-        quest_title: quest.title,
-        has_media: !!selectedMedia,
-        is_video: selectedMedia?.isVideo || false,
-        comfort_zone_rating: comfortRating,
-      });
-      setUserProperties({
-        [USER_PROPERTIES.LAST_ACTIVE]: new Date().toISOString(),
-      });
-    },
-    successMessage: t("Quest completed successfully!"),
-  });
-
-  const fadeIn = () => {
-    setModalVisible(true);
-    Animated.timing(fadeAnim, {
-      toValue: 1,
-      duration: 200,
-      useNativeDriver: true,
-    }).start();
-  };
-
-  const fadeOut = () => {
-    Animated.timing(fadeAnim, {
-      toValue: 0,
-      duration: 200,
-      useNativeDriver: true,
-    }).start(() => setModalVisible(false));
-  };
-
-  const onSubmit = async () => {
-    if (!user) {
-      alert(t("You must be logged in to submit a quest."));
-      return;
-    }
-
-    const placeholderText = t("Just completed today's quest!");
-    const textToSubmit = postText && postText.trim() ? postText : placeholderText;
-
-    await handleSubmit(
-      {
-        user_id: user.id,
-        quest_id: quest.id,
-        comfort_zone_rating: comfortRating,
-      },
-      textToSubmit
-    );
-  };
-
   return (
-    <>
-      <TouchableOpacity
-        style={[styles.ctaButton, completed && styles.completedButton]}
-        onPress={fadeIn}
-        disabled={completed || checking}
-      >
-        <View style={styles.ctaButtonContent}>
-          {completed && (
-            <MaterialIcons name="check-circle" size={20} color="#2D5016" style={styles.checkIcon} />
-          )}
-          <Text style={[styles.ctaButtonText, completed && styles.completedButtonText]}>
-            {completed ? t("Quest completed!") : t("Mark as complete")}
-          </Text>
-        </View>
-      </TouchableOpacity>
-
-      <Modal transparent visible={modalVisible} onRequestClose={fadeOut} statusBarTranslucent>
-        <Animated.View style={[StyleSheet.absoluteFillObject, styles.modalOverlay, { opacity: fadeAnim }]}>
-          <KeyboardAvoidingView
-            behavior={Platform.OS === "ios" ? "padding" : "height"}
-            style={styles.keyboardView}
-          >
-            <TouchableOpacity style={styles.modalContainer} activeOpacity={1} onPress={fadeOut}>
-              <View
-                style={styles.modalContent}
-                onStartShouldSetResponder={() => true}
-                onTouchEnd={(e) => e.stopPropagation()}
-              >
-                <View style={styles.modalHeader}>
-                  <Text style={styles.modalTitle}>{t("How did it go?")}</Text>
-                  <Text style={styles.modalSubtitle}>{t("Share how you completed the quest")}</Text>
-                  <TouchableOpacity onPress={fadeOut} style={styles.closeButton}>
-                    <MaterialIcons name="close" size={20} color={colors.light.text} />
-                  </TouchableOpacity>
-                </View>
-
-                <TouchableOpacity style={styles.mediaUploadButton} onPress={handleMediaUpload}>
-                  {selectedMedia ? (
-                    <View style={styles.mediaSelectedRow}>
-                      <Image
-                        source={{ uri: selectedMedia.thumbnailUri || selectedMedia.previewUrl }}
-                        style={styles.thumbnail}
-                        resizeMode="cover"
-                      />
-                      <View style={styles.mediaSelectedTextContainer}>
-                        <Text style={styles.mediaSelectedTitle}>
-                          {t(selectedMedia.isVideo ? "Video attached" : "Photo attached")}
-                        </Text>
-                        <Text style={styles.mediaSelectedSubtitle}>{t("Tap to change")}</Text>
-                      </View>
-                      <TouchableOpacity style={styles.removeButtonInline} onPress={handleRemoveMedia} disabled={isUploading}>
-                        <MaterialIcons name="close" size={18} color={colors.neutral.darkGrey} />
-                      </TouchableOpacity>
-                    </View>
-                  ) : (
-                    <View style={styles.mediaEmptyRow}>
-                      {isUploading ? (
-                        <Loader />
-                      ) : (
-                        <View style={styles.uploadMediaContainer}>
-                          <View style={styles.mediaIconsContainer}>
-                            <MaterialIcons name="image" size={18} color={colors.neutral.darkGrey} />
-                            <MaterialCommunityIcons name="video" size={18} color={colors.neutral.darkGrey} />
-                          </View>
-                          <Text style={styles.uploadButtonText}>{t("Add photo/video (optional)")}</Text>
-                        </View>
-                      )}
-                    </View>
-                  )}
-                </TouchableOpacity>
-
-                {uploadProgress !== null && (
-                  <View style={styles.uploadProgressContainer}>
-                    <Text style={styles.uploadProgressText}>
-                      {t("Uploading media...")} {Math.round(uploadProgress)}%
-                    </Text>
-                    <View style={styles.progressBarContainer}>
-                      <View style={[styles.progressBarFill, { width: `${uploadProgress}%` }]} />
-                    </View>
-                  </View>
-                )}
-
-                <TextInput
-                  style={styles.textInput}
-                  multiline
-                  scrollEnabled
-                  autoFocus
-                  placeholder={t("Just completed today's quest!")}
-                  placeholderTextColor="#999"
-                  onChangeText={setPostText}
-                />
-
-                <View style={styles.comfortRow}>
-                  <Text style={styles.comfortLabel}>{t("How far out of your comfort zone was this?")}</Text>
-                  <View style={styles.comfortButtons}>
-                    {[1, 2, 3, 4, 5].map((value) => (
-                      <TouchableOpacity
-                        key={value}
-                        style={[styles.comfortButton, comfortRating === value && styles.comfortButtonActive]}
-                        onPress={() => setComfortRating(value)}
-                      >
-                        <Text style={[styles.comfortButtonText, comfortRating === value && styles.comfortButtonTextActive]}>
-                          {value}
-                        </Text>
-                      </TouchableOpacity>
-                    ))}
-                  </View>
-                </View>
-
-                <TouchableOpacity
-                  style={[styles.submitButton, (isUploading || isSubmitting) && styles.submitButtonDisabled]}
-                  onPress={onSubmit}
-                  disabled={isUploading || isSubmitting}
-                >
-                  <Text style={styles.submitButtonText}>
-                    {t(isUploading ? "Uploading..." : isSubmitting ? "Submitting..." : "Submit")}
-                  </Text>
-                </TouchableOpacity>
-              </View>
-            </TouchableOpacity>
-          </KeyboardAvoidingView>
-        </Animated.View>
-      </Modal>
-    </>
+    <CompletionPostComposer
+      variant="quest"
+      completed={completed}
+      checkingCompletion={checking}
+      buildSubmitData={({ userId }) => ({
+        user_id: userId,
+        quest_id: quest.id,
+      })}
+      onCompleted={({ selectedMedia }) => {
+        queryClient.setQueryData(["quest-completion", quest.id, user?.id], true);
+        queryClient.invalidateQueries({ queryKey: ["home-posts"] });
+        captureEvent(SIDE_QUEST_EVENTS.DAILY_QUEST_COMPLETED, {
+          quest_id: quest.id,
+          quest_title: quest.title,
+          has_media: !!selectedMedia,
+          is_video: selectedMedia?.isVideo || false,
+        });
+        setUserProperties({
+          [USER_PROPERTIES.LAST_ACTIVE]: new Date().toISOString(),
+        });
+      }}
+    />
   );
 };
 
 const styles = StyleSheet.create({
-  comfortButton: {
-    alignItems: "center",
-    borderColor: colors.sideQuest.bgBorder,
-    borderRadius: 999,
-    borderWidth: 1,
-    height: 34,
-    justifyContent: "center",
-    width: 34,
-  },
-  comfortButtonActive: {
-    backgroundColor: colors.sideQuest.base,
-    borderColor: colors.sideQuest.base,
-  },
-  comfortButtonText: {
-    color: colors.sideQuest.text,
-    fontSize: 13,
-    fontWeight: "700",
-  },
-  comfortButtonTextActive: {
-    color: colors.neutral.white,
-  },
-  comfortButtons: {
-    flexDirection: "row",
-    gap: 8,
-  },
-  comfortLabel: {
-    color: colors.light.text,
-    fontSize: 13,
-    fontWeight: "600",
-    marginBottom: 10,
-  },
-  comfortRow: {
-    marginBottom: 18,
-  },
-  completedButton: {
-    backgroundColor: colors.light.easyGreen,
-  },
-  completedButtonText: {
-    color: "#2D5016",
-    fontWeight: "600",
-  },
-  checkIcon: {
-    marginRight: 8,
-  },
-  closeButton: {
-    position: "absolute",
-    right: 0,
-    top: 0,
-  },
-  ctaButton: {
-    alignItems: "center",
-    backgroundColor: colors.light.primary,
-    borderRadius: 10,
-    padding: 14,
-    width: "100%",
-  },
-  ctaButtonContent: {
-    alignItems: "center",
-    flexDirection: "row",
-    justifyContent: "center",
-  },
-  ctaButtonText: {
-    color: colors.neutral.white,
-    fontSize: 16,
-    fontWeight: "700",
-  },
   heroDetailHeader: {
     alignItems: "center",
     flexDirection: "row",
@@ -498,131 +249,7 @@ const styles = StyleSheet.create({
     fontWeight: "800",
     lineHeight: 31,
   },
-  keyboardView: {
-    flex: 1,
-  },
-  mediaEmptyRow: {
-    minHeight: 72,
-    justifyContent: "center",
-  },
-  mediaIconsContainer: {
-    flexDirection: "row",
-    gap: 8,
-  },
-  mediaSelectedRow: {
-    alignItems: "center",
-    flexDirection: "row",
-  },
-  mediaSelectedSubtitle: {
-    color: colors.light.lightText,
-    fontSize: 13,
-  },
-  mediaSelectedTextContainer: {
-    flex: 1,
-    marginLeft: 12,
-  },
-  mediaSelectedTitle: {
-    color: colors.light.text,
-    fontSize: 14,
-    fontWeight: "600",
-  },
-  mediaUploadButton: {
-    borderColor: colors.neutral.grey2,
-    borderRadius: 12,
-    borderWidth: 1,
-    marginBottom: 14,
-    padding: 14,
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: "flex-end",
-  },
-  modalContent: {
-    backgroundColor: colors.neutral.white,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    padding: 20,
-  },
-  modalHeader: {
-    marginBottom: 14,
-    position: "relative",
-  },
-  modalOverlay: {
-    backgroundColor: "rgba(0,0,0,0.4)",
-  },
-  modalSubtitle: {
-    color: colors.light.lightText,
-    fontSize: 14,
-    marginTop: 4,
-  },
-  modalTitle: {
-    color: colors.light.text,
-    fontSize: 20,
-    fontWeight: "700",
-  },
-  progressBarContainer: {
-    backgroundColor: colors.neutral.grey2,
-    borderRadius: 999,
-    height: 6,
-    overflow: "hidden",
-  },
-  progressBarFill: {
-    backgroundColor: colors.sideQuest.base,
-    height: "100%",
-  },
   questPage: {
     paddingHorizontal: 4,
-  },
-  removeButtonInline: {
-    padding: 4,
-  },
-  submitButton: {
-    alignItems: "center",
-    backgroundColor: colors.sideQuest.base,
-    borderRadius: 12,
-    padding: 14,
-  },
-  submitButtonDisabled: {
-    opacity: 0.5,
-  },
-  submitButtonText: {
-    color: colors.neutral.white,
-    fontSize: 15,
-    fontWeight: "700",
-  },
-  textInput: {
-    borderColor: colors.neutral.grey2,
-    borderRadius: 12,
-    borderWidth: 1,
-    color: colors.light.text,
-    fontSize: 14,
-    marginBottom: 14,
-    maxHeight: 120,
-    minHeight: 110,
-    padding: 14,
-    textAlignVertical: "top",
-  },
-  thumbnail: {
-    borderRadius: 10,
-    height: 56,
-    width: 56,
-  },
-  uploadButtonText: {
-    color: colors.light.text,
-    fontSize: 14,
-    fontWeight: "600",
-  },
-  uploadMediaContainer: {
-    alignItems: "center",
-    flexDirection: "row",
-    gap: 10,
-  },
-  uploadProgressContainer: {
-    marginBottom: 14,
-  },
-  uploadProgressText: {
-    color: colors.light.lightText,
-    fontSize: 12,
-    marginBottom: 6,
   },
 });
